Split monolithic src/read.c into .h and .c files
New items: box, decodeinput, decoderdata, decoderitem, layout, meta,
           parse, sample
Everything is kept in the same relative order it was in read.c.
GNU 11 static Release make in ~7.7s locally (+8%).
Binary size before/after:
  avifdec    9258424 -> 9259840 (+0.02%)
  libavif.a   217540 ->  233556 (+7.36%)
Clang 11 shared Release make -j 10 in ~2.3s locally (-2%).
Binary size before/after:
  libavif.so 8553840 -> 8551136 (-0.03%)
