+++ src/decoderdata.c
@@ -1,0 +1,109 @@
+// Copyright 2019 Joe Drago. All rights reserved.
+// SPDX-License-Identifier: BSD-2-Clause
+
+#include "avif/avif.h"
+#include "avif/internal.h"
+#include "decodeinput.h"
+#include "decoderdata.h"
+#include "meta.h"
+
+#include <assert.h>
+#include <inttypes.h>
+#include <limits.h>
+#include <string.h>
+
+avifDecoderData * avifDecoderDataCreate()
+{
+    avifDecoderData * data = (avifDecoderData *)avifAlloc(sizeof(avifDecoderData));
+    memset(data, 0, sizeof(avifDecoderData));
+    data->meta = avifMetaCreate();
+    avifArrayCreate(&data->tracks, sizeof(avifTrack), 2);
+    avifArrayCreate(&data->tiles, sizeof(avifTile), 8);
+    return data;
+}
+
+void avifDecoderDataResetCodec(avifDecoderData * data)
+{
+    for (unsigned int i = 0; i < data->tiles.count; ++i) {
+        avifTile * tile = &data->tiles.tile[i];
+        if (tile->image) {
+            avifImageFreePlanes(tile->image, AVIF_PLANES_ALL); // forget any pointers into codec image buffers
+        }
+        if (tile->codec) {
+            avifCodecDestroy(tile->codec);
+            tile->codec = NULL;
+        }
+    }
+}
+
+avifTile * avifDecoderDataCreateTile(avifDecoderData * data, uint32_t width, uint32_t height, uint8_t operatingPoint)
+{
+    avifTile * tile = (avifTile *)avifArrayPushPtr(&data->tiles);
+    tile->image = avifImageCreateEmpty();
+    tile->input = avifCodecDecodeInputCreate();
+    tile->width = width;
+    tile->height = height;
+    tile->operatingPoint = operatingPoint;
+    return tile;
+}
+
+avifTrack * avifDecoderDataCreateTrack(avifDecoderData * data)
+{
+    avifTrack * track = (avifTrack *)avifArrayPushPtr(&data->tracks);
+    track->meta = avifMetaCreate();
+    return track;
+}
+
+void avifDecoderDataClearTiles(avifDecoderData * data)
+{
+    for (unsigned int i = 0; i < data->tiles.count; ++i) {
+        avifTile * tile = &data->tiles.tile[i];
+        if (tile->input) {
+            avifCodecDecodeInputDestroy(tile->input);
+            tile->input = NULL;
+        }
+        if (tile->codec) {
+            avifCodecDestroy(tile->codec);
+            tile->codec = NULL;
+        }
+        if (tile->image) {
+            avifImageDestroy(tile->image);
+            tile->image = NULL;
+        }
+    }
+    data->tiles.count = 0;
+    data->colorTileCount = 0;
+    data->alphaTileCount = 0;
+}
+
+void avifDecoderDataDestroy(avifDecoderData * data)
+{
+    avifMetaDestroy(data->meta);
+    for (uint32_t i = 0; i < data->tracks.count; ++i) {
+        avifTrack * track = &data->tracks.track[i];
+        if (track->sampleTable) {
+            avifSampleTableDestroy(track->sampleTable);
+        }
+        if (track->meta) {
+            avifMetaDestroy(track->meta);
+        }
+    }
+    avifArrayDestroy(&data->tracks);
+    avifDecoderDataClearTiles(data);
+    avifArrayDestroy(&data->tiles);
+    avifFree(data);
+}
+
+void avifDecoderCleanup(avifDecoder * decoder)
+{
+    if (decoder->data) {
+        avifDecoderDataDestroy(decoder->data);
+        decoder->data = NULL;
+    }
+
+    if (decoder->image) {
+        avifImageDestroy(decoder->image);
+        decoder->image = NULL;
+    }
+    avifDiagnosticsClearError(&decoder->diag);
+}
