+++ src/sample.c
@@ -1,0 +1,142 @@
+// Copyright 2019 Joe Drago. All rights reserved.
+// SPDX-License-Identifier: BSD-2-Clause
+
+#include "avif/internal.h"
+#include "decoderdata.h"
+#include "decoderitem.h"
+#include "meta.h"
+#include "sample.h"
+
+#include <assert.h>
+#include <inttypes.h>
+#include <limits.h>
+#include <string.h>
+
+avifSampleTable * avifSampleTableCreate()
+{
+    avifSampleTable * sampleTable = (avifSampleTable *)avifAlloc(sizeof(avifSampleTable));
+    memset(sampleTable, 0, sizeof(avifSampleTable));
+    avifArrayCreate(&sampleTable->chunks, sizeof(avifSampleTableChunk), 16);
+    avifArrayCreate(&sampleTable->sampleDescriptions, sizeof(avifSampleDescription), 2);
+    avifArrayCreate(&sampleTable->sampleToChunks, sizeof(avifSampleTableSampleToChunk), 16);
+    avifArrayCreate(&sampleTable->sampleSizes, sizeof(avifSampleTableSampleSize), 16);
+    avifArrayCreate(&sampleTable->timeToSamples, sizeof(avifSampleTableTimeToSample), 16);
+    avifArrayCreate(&sampleTable->syncSamples, sizeof(avifSyncSample), 16);
+    return sampleTable;
+}
+
+void avifSampleTableDestroy(avifSampleTable * sampleTable)
+{
+    avifArrayDestroy(&sampleTable->chunks);
+    for (uint32_t i = 0; i < sampleTable->sampleDescriptions.count; ++i) {
+        avifSampleDescription * description = &sampleTable->sampleDescriptions.description[i];
+        avifArrayDestroy(&description->properties);
+    }
+    avifArrayDestroy(&sampleTable->sampleDescriptions);
+    avifArrayDestroy(&sampleTable->sampleToChunks);
+    avifArrayDestroy(&sampleTable->sampleSizes);
+    avifArrayDestroy(&sampleTable->timeToSamples);
+    avifArrayDestroy(&sampleTable->syncSamples);
+    avifFree(sampleTable);
+}
+
+uint32_t avifSampleTableGetImageDelta(const avifSampleTable * sampleTable, int imageIndex)
+{
+    int maxSampleIndex = 0;
+    for (uint32_t i = 0; i < sampleTable->timeToSamples.count; ++i) {
+        const avifSampleTableTimeToSample * timeToSample = &sampleTable->timeToSamples.timeToSample[i];
+        maxSampleIndex += timeToSample->sampleCount;
+        if ((imageIndex < maxSampleIndex) || (i == (sampleTable->timeToSamples.count - 1))) {
+            return timeToSample->sampleDelta;
+        }
+    }
+
+    // TODO: fail here?
+    return 1;
+}
+
+avifBool avifSampleTableHasFormat(const avifSampleTable * sampleTable, const char * format)
+{
+    for (uint32_t i = 0; i < sampleTable->sampleDescriptions.count; ++i) {
+        if (!memcmp(sampleTable->sampleDescriptions.description[i].format, format, 4)) {
+            return AVIF_TRUE;
+        }
+    }
+    return AVIF_FALSE;
+}
+
+const avifPropertyArray * avifSampleTableGetProperties(const avifSampleTable * sampleTable)
+{
+    for (uint32_t i = 0; i < sampleTable->sampleDescriptions.count; ++i) {
+        const avifSampleDescription * description = &sampleTable->sampleDescriptions.description[i];
+        if (!memcmp(description->format, "av01", 4)) {
+            return &description->properties;
+        }
+    }
+    return NULL;
+}
+
+uint32_t avifGetSampleCountOfChunk(const avifSampleTableSampleToChunkArray * sampleToChunks, uint32_t chunkIndex)
+{
+    uint32_t sampleCount = 0;
+    for (int sampleToChunkIndex = sampleToChunks->count - 1; sampleToChunkIndex >= 0; --sampleToChunkIndex) {
+        const avifSampleTableSampleToChunk * sampleToChunk = &sampleToChunks->sampleToChunk[sampleToChunkIndex];
+        if (sampleToChunk->firstChunk <= (chunkIndex + 1)) {
+            sampleCount = sampleToChunk->samplesPerChunk;
+            break;
+        }
+    }
+    return sampleCount;
+}
+
+avifResult avifDecoderPrepareSample(avifDecoder * decoder, avifDecodeSample * sample, size_t partialByteCount)
+{
+    if (!sample->data.size || sample->partialData) {
+        // This sample hasn't been read from IO or had its extents fully merged yet.
+
+        size_t bytesToRead = sample->size;
+        if (partialByteCount && (bytesToRead > partialByteCount)) {
+            bytesToRead = partialByteCount;
+        }
+
+        if (sample->itemID) {
+            // The data comes from an item. Let avifDecoderItemRead() do the heavy lifting.
+
+            avifDecoderItem * item = avifMetaFindItem(decoder->data->meta, sample->itemID);
+            avifROData itemContents;
+            avifResult readResult = avifDecoderItemRead(item, decoder->io, &itemContents, sample->offset, bytesToRead, &decoder->diag);
+            if (readResult != AVIF_RESULT_OK) {
+                return readResult;
+            }
+
+            // avifDecoderItemRead is guaranteed to already be persisted by either the underlying IO
+            // or by mergedExtents; just reuse the buffer here.
+            memcpy(&sample->data, &itemContents, sizeof(avifROData));
+            sample->ownsData = AVIF_FALSE;
+            sample->partialData = item->partialMergedExtents;
+        } else {
+            // The data likely comes from a sample table. Pull the sample and make a copy if necessary.
+
+            avifROData sampleContents;
+            if ((decoder->io->sizeHint > 0) && (sample->offset > decoder->io->sizeHint)) {
+                return AVIF_RESULT_BMFF_PARSE_FAILED;
+            }
+            avifResult readResult = decoder->io->read(decoder->io, 0, sample->offset, bytesToRead, &sampleContents);
+            if (readResult != AVIF_RESULT_OK) {
+                return readResult;
+            }
+            if (sampleContents.size != bytesToRead) {
+                return AVIF_RESULT_TRUNCATED_DATA;
+            }
+
+            sample->ownsData = !decoder->io->persistent;
+            sample->partialData = (bytesToRead != sample->size);
+            if (decoder->io->persistent) {
+                memcpy(&sample->data, &sampleContents, sizeof(avifROData));
+            } else {
+                avifRWDataSet((avifRWData *)&sample->data, sampleContents.data, sampleContents.size);
+            }
+        }
+    }
+    return AVIF_RESULT_OK;
+}
