+++ src/sample.h
@@ -1,0 +1,81 @@
+// Copyright 2019 Joe Drago. All rights reserved.
+// SPDX-License-Identifier: BSD-2-Clause
+
+#ifndef SRC_SAMPLE_H_
+#define SRC_SAMPLE_H_
+
+#include "avif/internal.h"
+#include "box.h"
+
+#include <assert.h>
+#include <inttypes.h>
+#include <limits.h>
+#include <string.h>
+
+// ---------------------------------------------------------------------------
+// avifTrack
+
+typedef struct avifSampleTableChunk
+{
+    uint64_t offset;
+} avifSampleTableChunk;
+AVIF_ARRAY_DECLARE(avifSampleTableChunkArray, avifSampleTableChunk, chunk);
+
+typedef struct avifSampleTableSampleToChunk
+{
+    uint32_t firstChunk;
+    uint32_t samplesPerChunk;
+    uint32_t sampleDescriptionIndex;
+} avifSampleTableSampleToChunk;
+AVIF_ARRAY_DECLARE(avifSampleTableSampleToChunkArray, avifSampleTableSampleToChunk, sampleToChunk);
+
+typedef struct avifSampleTableSampleSize
+{
+    uint32_t size;
+} avifSampleTableSampleSize;
+AVIF_ARRAY_DECLARE(avifSampleTableSampleSizeArray, avifSampleTableSampleSize, sampleSize);
+
+typedef struct avifSampleTableTimeToSample
+{
+    uint32_t sampleCount;
+    uint32_t sampleDelta;
+} avifSampleTableTimeToSample;
+AVIF_ARRAY_DECLARE(avifSampleTableTimeToSampleArray, avifSampleTableTimeToSample, timeToSample);
+
+typedef struct avifSyncSample
+{
+    uint32_t sampleNumber;
+} avifSyncSample;
+AVIF_ARRAY_DECLARE(avifSyncSampleArray, avifSyncSample, syncSample);
+
+typedef struct avifSampleDescription
+{
+    uint8_t format[4];
+    avifPropertyArray properties;
+} avifSampleDescription;
+AVIF_ARRAY_DECLARE(avifSampleDescriptionArray, avifSampleDescription, description);
+
+typedef struct avifSampleTable
+{
+    avifSampleTableChunkArray chunks;
+    avifSampleDescriptionArray sampleDescriptions;
+    avifSampleTableSampleToChunkArray sampleToChunks;
+    avifSampleTableSampleSizeArray sampleSizes;
+    avifSampleTableTimeToSampleArray timeToSamples;
+    avifSyncSampleArray syncSamples;
+    uint32_t allSamplesSize; // If this is non-zero, sampleSizes will be empty and all samples will be this size
+} avifSampleTable;
+
+avifSampleTable * avifSampleTableCreate(void);
+void avifSampleTableDestroy(avifSampleTable * sampleTable);
+uint32_t avifSampleTableGetImageDelta(const avifSampleTable * sampleTable, int imageIndex);
+avifBool avifSampleTableHasFormat(const avifSampleTable * sampleTable, const char * format);
+
+const avifPropertyArray * avifSampleTableGetProperties(const avifSampleTable * sampleTable);
+
+// Returns how many samples are in the chunk.
+uint32_t avifGetSampleCountOfChunk(const avifSampleTableSampleToChunkArray * sampleToChunks, uint32_t chunkIndex);
+
+avifResult avifDecoderPrepareSample(avifDecoder * decoder, avifDecodeSample * sample, size_t partialByteCount);
+
+#endif  // SRC_SAMPLE_H_
